+++ backend/src/share/dto/myShare.dto.ts
@@ -2,6 +2,7 @@ import { Expose, plainToClass, Type } from "class-transformer";
 import { ShareDTO } from "./share.dto";
 import { FileDTO } from "../../file/dto/file.dto";
 import { OmitType } from "@nestjs/swagger";
+import { MyShareSecurityDTO } from "./myShareSecurity.dto";
 
 export class MyShareDTO extends OmitType(ShareDTO, [
   "files",
@@ -20,6 +21,9 @@ export class MyShareDTO extends OmitType(ShareDTO, [
   @Expose()
   @Type(() => OmitType(FileDTO, ["share", "from"] as const))
   files: Omit<FileDTO, "share" | "from">[];
+
+  @Expose()
+  security?: MyShareSecurityDTO;
 
   from(partial: Partial<MyShareDTO>) {
     return plainToClass(MyShareDTO, partial, { excludeExtraneousValues: true });