+++ frontend/src/utils/date.util.ts
@@ -1,4 +1,5 @@
 import moment from "moment";
+import { Timespan } from "../types/timespan.type";
 
 export const getExpirationPreview = (
   messages: {
@@ -29,4 +30,15 @@ export const getExpirationPreview = (
     "{expiration}",
     moment(expirationDate).format("LLL"),
   );
+};
+
+export const timespanToString = (timespan: Timespan) => {
+  return `${timespan.value} ${timespan.unit}`;
+}
+
+export const stringToTimespan = (value: string): Timespan => {
+  return {
+    value: parseInt(value.split(" ")[0]),
+    unit: value.split(" ")[1],
+  } as Timespan;
 };